+++ src/edu/alfonso_depaz/Evidencia3/process/metodos/jugadores/Metodo_RegistroDatos_Jugador1_V1.java
@@ -1,15 +1,13 @@
 /**
  * Clase que sobrescribe los métodos para registrar jugadores y caracteres del usuario.
- * Esta variante de registro del jugador 1 utiliza salidas "System.out.println" para solicitar
- * al usuario los datos requeridos.
  */
-package edu.alfonso_depaz.MiJuego1.process.metodos.jugadores;
+package edu.alfonso_depaz.Evidencia3.process.metodos.jugadores;
 
-import edu.alfonso_depaz.MiJuego1.ui.Idioma;
+import edu.alfonso_depaz.Evidencia3.ui.Idioma;
 
 import java.util.Scanner;
 
-public class Metodo_RegistroDatos_Jugador2_V1 extends Metodos_Jugadores {
+public class Metodo_RegistroDatos_Jugador1_V1 extends Metodos_Jugadores {
 
     /**
      * Método para guardar el nombre del jugador.
@@ -19,12 +17,11 @@ public class Metodo_RegistroDatos_Jugador2_V1 extends Metodos_Jugadores {
      */
     @Override
     public String Guardar_NombresJugadores_V1(String jugador, Idioma lenguaje) {
-        // Se invoca scanner para leer salidas del usuario
         Scanner scanner = new Scanner(System.in);
-        // Se imprime mensaje
-        System.out.println(lenguaje.get_Entrada_Nombre_Jg_Dos());
-        // Se registra el nombre
+        // Se imprime mensaje para registrar el jugador 1
+        System.out.println(lenguaje.get_Entrada_Nombre_Jg_Uno());
         jugador = scanner.nextLine();
+        // Se ejecuta el método leyéndolo y regresando su "valor"
         return jugador;
     }
 
@@ -41,7 +38,7 @@ public class Metodo_RegistroDatos_Jugador2_V1 extends Metodos_Jugadores {
         // Se imprime mensaje de la lista de caracteres
         System.out.println(lenguaje.mostrar_Salida_CaracteresDeTicTacToe());
         // Se imprime mensaje para registrar el jugador
-        System.out.println(lenguaje.get_Entrada_CaracteresDeTicTacToe_Jg_Dos());
+        System.out.println(lenguaje.get_Entrada_CaracteresDeTicTacToe_Jg_Uno());
         // Se registra la opción
         opcion = scanner.nextLine();
