+++ src/edu/alfonso_depaz/Evidencia3/ui/English.java
@@ -1,42 +1,42 @@
-package edu.alfonso_depaz.MiJuego1.ui;
+package edu.alfonso_depaz.Evidencia3.ui;
 
 /**
- * Clase que define los mensajes en japonés para el juego.
+ * Clase que define los mensajes en inglés para el juego.
  */
-public class Nihongo extends Idioma{
+public class English extends Idioma{
 
     /**
-     * Mensaje de bienvenida en japonés.
+     * Mensaje de bienvenida en inglés.
      * @return mensaje de bienvenida
      */
     @Override
     public String get_Bienvenida() {
-        return """
-               TIC TAC TOE ゲームへようこそ
-               """;
+        return "Welcome";
     }
 
     /**
-     * Mensaje que solicita al usuario seleccionar el modo de juego en japonés.
+     * Mensaje que solicita al usuario seleccionar el modo de juego en inglés.
      * @return mensaje de solicitud de modo de juego
      */
     @Override
     public String get_Entrada_ModoDeJuego() {
         return """
-               ///////////////////////////////////////////////////////////
-               
-               ゲームモード？
-               
-               1. プレイヤー対プレイヤー
-               
-               2. プレイヤー対 CPU
-               
-               ///////////////////////////////////////////////////////////
-               """;
+                ///////////////////////////////////////////////////////////
+                
+                Choose your game mode...
+                
+                1. Player vs Player
+                
+                2. Player vs CPU
+                
+                3. Show hall of fame
+                
+                ///////////////////////////////////////////////////////////
+                """;
     }
 
     /**
-     * Muestra la representación de caracteres del juego de Tic Tac Toe en japonés.
+     * Muestra la representación de caracteres del juego de Tic Tac Toe en inglés.
      * @return representación de caracteres del juego
      */
     @Override
@@ -44,8 +44,7 @@ public class Nihongo extends Idioma{
         return """
                 ///////////////////////////////////////////////////////////
                 
-                これらは
-                チックタックトゥ」のボックスをマークする文字です"
+                These are the characters for the "Tic Tac Toe"
                 
                 A. X
                 B. O
@@ -63,191 +62,191 @@ public class Nihongo extends Idioma{
     }
 
     /**
-     * Mensaje que solicita al usuario ingresar los caracteres para el jugador 1 en japonés.
+     * Mensaje que solicita al usuario ingresar los caracteres para el jugador 1 en inglés.
      * @return mensaje de solicitud de caracteres para el jugador 1
      */
     @Override
     public String get_Entrada_CaracteresDeTicTacToe_Jg_Uno() {
-        return "「プレイヤー 1 のキャラクター:」:";
+        return "Characters for the player 1:";
     }
 
     /**
-     * Mensaje que solicita al usuario ingresar los caracteres para el jugador 2 en japonés.
+     * Mensaje que solicita al usuario ingresar los caracteres para el jugador 2 en inglés.
      * @return mensaje de solicitud de caracteres para el jugador 2
      */
     @Override
     public String get_Entrada_CaracteresDeTicTacToe_Jg_Dos() {
-        return "「プレイヤー 2 のキャラクター:」:";
+        return "Characters for the player 2:";
     }
 
     /**
-     * Mensaje que solicita al usuario ingresar el nombre del jugador 1 en japonés.
+     * Mensaje que solicita al usuario ingresar el nombre del jugador 1 en inglés.
      * @return mensaje de solicitud de nombre del jugador 1
      */
     @Override
     public String get_Entrada_Nombre_Jg_Uno() {
-        return "「プレイヤー１の名前は？」";
+        return "Name of the player 1?";
     }
 
     /**
-     * Mensaje que solicita al usuario ingresar el nombre del jugador 2 en japonés.
+     * Mensaje que solicita al usuario ingresar el nombre del jugador 2 en inglés.
      * @return mensaje de solicitud de nombre del jugador 2
      */
     @Override
     public String get_Entrada_Nombre_Jg_Dos() {
-        return "「プレイヤー2の名前は？」";
+        return "Name of the player 2";
     }
 
     /**
-     * Mensaje de error de caracter inválido en japonés.
+     * Mensaje de error de caracter inválido en inglés.
      * @return mensaje de error de caracter inválido
      */
     @Override
     public String get_Salida_Error_CaracterInvalido() {
-        return "「無効な文字です」";
+        return "Invalid Character";
     }
 
     /**
-     * Mensaje de bienvenida al modo de juego Jugador vs Jugador en japonés.
+     * Mensaje de bienvenida al modo de juego Jugador vs Jugador en inglés.
      * @return mensaje de bienvenida al modo de juego Jugador vs Jugador
      */
     @Override
     public String get_Salida_Bienvenida_ModoDeJuego_JgVSJg() {
-        return "「ゲームモード: プレイヤー vs プレイヤー...」";
+        return "Game mode: Player Vs Player";
     }
 
     /**
-     * Mensaje de bienvenida al modo de juego Jugador vs CPU en japonés.
+     * Mensaje de bienvenida al modo de juego Jugador vs CPU en inglés.
      * @return mensaje de bienvenida al modo de juego Jugador vs CPU
      */
     @Override
     public String get_Salida_Bienvenida_ModoDeJuego_JgVSCPU() {
-        return "「ゲームモード: プレイヤー vs CPU...」";
+        return "Player Vs CPU";
     }
 
     /**
-     * Mensaje de confirmación de caracter registrado en japonés.
+     * Mensaje de confirmación de caracter registrado en inglés.
      * @return mensaje de confirmación de caracter registrado
      */
     @Override
     public String get_Salida_Confirmación_CaracterRegistrado() {
-        return "「登録キャラクター：」";
+        return "Character registered: ";
     }
 
     /**
-     * Mensaje de inicio de juego en japonés.
+     * Mensaje de inicio de juego en inglés.
      * @return mensaje de inicio de juego
      */
     @Override
     public String get_Salida_InicioJuego() {
-        return "「ゲームの始まり」";
+        return "Starting game";
     }
 
     /**
-     * Mensaje para finalizar el programa en japonés.
+     * Mensaje para finalizar el programa en inglés.
      * @return mensaje para finalizar el programa
      */
     @Override
     public String get_Salida_FinalizarPrograma() {
-        return "「プログラムを終了するには、いつでもFINを押してください。」";
+        return "To exit, use OUT";
     }
 
     /**
-     * Mensaje de finalización de juego en japonés.
+     * Mensaje de finalización de juego en inglés.
      * @return mensaje de finalización de juego
      */
     @Override
     public String get_Salida_FinalizandoJuego() {
-        return "「ゲームから離れる」";
+        return ";)";
     }
 
     /**
-     * Mensaje para regresar al menú en japonés.
+     * Mensaje para regresar al menú en inglés.
      * @return mensaje para regresar al menú
      */
     @Override
     public String get_Salida_RegresarMenu() {
-        return "「メニューに戻る」";
+        return "Returning to menu";
     }
 
     /**
-     * Mensaje de error de nombres de jugadores iguales en japonés.
+     * Mensaje de error de nombres de jugadores iguales en inglés.
      * @return mensaje de error de nombres de jugadores iguales
      */
     @Override
     public String get_Error_JugadoresIguales() {
-        return "「同じプレイヤー名です。同じ名前を入力しないでください。」";
+        return "Same Player Names, please do not enter the same names";
     }
 
     /**
-     * Mensaje para cambiar el nombre de un jugador en japonés.
+     * Mensaje para cambiar el nombre de un jugador en inglés.
      * @return mensaje para cambiar el nombre de un jugador
      */
     @Override
     public String get_Entrada_CambiarNombreJugadorV() {
-        return "「どの選手を変更しますか? 1/2」";
+        return "Which player do you want to change? 1/2";
     }
 
     /**
-     * Mensaje de error de caracteres de jugadores iguales en japonés.
+     * Mensaje de error de caracteres de jugadores iguales en inglés.
      * @return mensaje de error de caracteres de jugadores iguales
      */
     @Override
     public String get_Error_CaracteresIguales() {
-        return "「同じプレイヤーのシンボルです。別のシンボルを入力してください」";
+        return "Same player characters, please enter different characters";
     }
 
     /**
-     * Mensaje para cambiar el caracter de un jugador en japonés.
+     * Mensaje para cambiar el caracter de un jugador en inglés.
      * @return mensaje para cambiar el caracter de un jugador
      */
     @Override
     public String get_Entrada_CambiarCaracterJugador() {
-        return "「どのプレイヤーキャラクターを変更しますか? 1/2」";
+        return "Which player character do you want to change? 1/2";
     }
 
     /**
-     * Mensaje de empate en japonés.
+     * Mensaje de empate en inglés.
      * @return mensaje de empate
      */
     @Override
     public String get_Salida_Empate() {
-        return "「ネクタイ」";
+        return "Draw";
     }
 
     /**
-     * Mensaje de victoria en japonés.
+     * Mensaje de victoria en inglés.
      * @return mensaje de victoria
      */
     @Override
     public String get_Salida_Ganador() {
-        return "「勝て」！";
+        return "Won!";
     }
 
     /**
-     * Mensaje para jugar otra vez en japonés.
-     * @return mensaje para jugar otra vez
+     * Mensaje que solicita al usuario si desea jugar otra vez en inglés.
+     * @return mensaje de solicitud de jugar otra vez
      */
     @Override
     public String get_Salida_JugarOtraVez() {
         return """
-                また遊びたい？
-                1. うん
-                2. いいえ
+                Do you want to play again?
+                1. Yes
+                2. Nop
                 """;
     }
 
     /**
-     * Mensaje para indicar el turno del jugador en japonés.
-     * @return mensaje para indicar el turno del jugador
+     * Mensaje que solicita al usuario ingresar filas y columnas en inglés.
+     * @return mensaje de solicitud de filas y columnas
      */
     @Override
     public String get_Salida_TurnoDe() {
-        return "「ターン・オブ・ザ・:」: ";
+        return "Turn of: ";
     }
 
     /**
-     * Mensaje para solicitar al usuario que escoja filas y columnas en japonés.
+     * Mensaje que solicita al usuario ingresar filas y columnas en inglés.
      * @return mensaje de solicitud de filas y columnas
      */
     @Override
@@ -255,63 +254,63 @@ public class Nihongo extends Idioma{
         return """
                 //////////////////////////////////////////////
                 
-                "次に、行と列を置きます..."
+                Next, put your rows and columns...
                 
                 //////////////////////////////////////////////
                 """;
     }
 
     /**
-     * Mensaje para solicitar al usuario ingresar las filas en japonés.
+     * Mensaje que solicita al usuario ingresar las filas en inglés.
      * @return mensaje de solicitud de filas
      */
     @Override
     public String get_Entrada_ObtenerFilas() {
-        return "行？";
+        return "Rows?";
     }
 
     /**
-     * Mensaje para solicitar al usuario ingresar las columnas en japonés.
+     * Mensaje que solicita al usuario ingresar las columnas en inglés.
      * @return mensaje de solicitud de columnas
      */
     @Override
     public String get_Entrada_ObtenerColumnas() {
-        return "カラム？";
+        return "Columns?";
     }
 
     /**
-     * Mensaje de celda ocupada en japonés.
+     * Mensaje de celda ocupada en inglés.
      * @return mensaje de celda ocupada
      */
     @Override
     public String get_Salida_CeldaOcupada() {
-        return "「独房は占拠されています」";
+        return "The cell is occupied";
     }
 
     /**
-     * Mensaje que indica que el Salón de la Fama está vacío en japonés.
+     * Mensaje que indica que el Salón de la Fama está vacío en inglés.
      * @return mensaje de Salón de la Fama vacío
      */
     @Override
     public String get_Salida_SalonFamaVacio() {
-        return "「殿堂は空だ」";
+        return "The Hall of Fame is empty";
     }
 
     /**
-     * Mensaje que muestra el Salón de la Fama en japonés.
+     * Mensaje que muestra el Salón de la Fama en inglés.
      * @return mensaje del Salón de la Fama
      */
     @Override
     public String get_Salida_SalonFama() {
-        return "殿堂...";
+        return "Hall of Fame...";
     }
 
     /**
-     * Mensaje que muestra el número de victorias en japonés.
+     * Mensaje que muestra el número de victorias en inglés.
      * @return mensaje del número de victorias
      */
     @Override
     public String get_Salida_Victorias() {
-        return "勝利";
+        return " Victories";
     }
 }
